enhance AI response (#2)
* feat: extract texts from links

* together AI working

* together AI better ver

* together AI

* folderize

* organized v2

* organized v3

* update readme

* remove binary file

* test after suggestion

* test okay

* test

+++ libraries/together_ai/models/AI.go
@@ -47,7 +47,7 @@ type MessageEx struct {
 }
 
 type Usage struct {
-	PromptTokens     int `json:"prompt_tokens"`
-	TotalTokens      int `json:"total_tokens"`
-	CompletionTokens int `json:"completion_tokens"`
+	PromptTokens     float64 `json:"prompt_tokens"`
+	TotalTokens      float64 `json:"total_tokens"`
+	CompletionTokens float64 `json:"completion_tokens"`
 }
